storage: move storage server related helpers from hsm
HSM module is huge and contains various function related only to some
other module. Such functions are e.g. helpers for connecting to storage
servers. Move these functions partially into storageServer module and
those related only to iscsi connection to iscsi module.

Bug-Url: https://bugzilla.redhat.com/1787192
